requirement.txt
annotated-types==0.7.0
anyio==4.6.0
argcomplete==3.5.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.29.1
decli==0.6.2
distlib==0.3.8
dnspython==2.6.1
fastapi==0.115.0
filelock==3.16.1
h11==0.14.0
identify==2.6.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
motor==3.6.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.3.6
pre-commit==3.8.0
prompt-toolkit==3.0.36
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dotenv==1.0.1
PyYAML==6.0.2
questionary==2.0.1
sniffio==1.3.1
starlette==0.38.6
termcolor==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.30.6
virtualenv==20.26.6
wcwidth==0.2.13
